+++ plugins/data-management/src/HierarchicalTrackSelectorWidget/components/DefaultFolderDialog.tsx
@@ -5,10 +5,10 @@ import { getSession } from '@jbrowse/core/util'
 import { DialogContent } from '@mui/material'
 import { observer } from 'mobx-react'
 
-import { facetedStateTreeF } from '../facetedModel.ts'
-import FacetedSelector from './faceted/FacetedSelector.tsx'
+import { facetedStateTreeF } from '../../FacetedSelector/facetedModel.ts'
+import FacetedSelector from '../../FacetedSelector/components/FacetedSelector.tsx'
 
-import type { FacetedModel } from '../facetedModel.ts'
+import type { FacetedModel } from '../../FacetedSelector/facetedModel.ts'
 import type { HierarchicalTrackSelectorModel } from '../model.ts'
 import type { TreeNode, TreeTrackNode } from '../types.ts'
